# sutra/contracts/sutra.openapi.yaml
openapi: 3.0.3
info:
  title: sutra
  description: The single API surface — CLI, web UI, and external agents all speak this contract.
  version: 0.2.0
paths:
...
  /projects/import:
    post:
      operationId: importProject
      summary: Import a project export
      description: >-
        Reconstructs a project from a previous export with UUIDs preserved
        (AC-import-round-trip). Identities referenced by the export's records are
        recreated from the export itself — never invented — so every actor and
        assignee resolves after import. Exported records are canonical
        serializations — read-time decorations like feed_watermark are never
        part of them, and a payload carrying one is rejected as malformed — every field, including server-stamped ones (e.g.
        Comment.review_revision), is preserved on import verbatim, not
        recomputed. Import deliberately targets empty servers: rejected as a
        whole, naming the conflict, if any UUID already exists on this
        server (AC-import-collision) — any pre-existing UUID is a
        collision, including identity ids; no partial imports. Hierarchies
        are validated against the completion invariant: a complete issue
        with any descendant in open, in-progress, or blocked — at any depth,
        including beneath deferred children — rejects the import whole as
        malformed (AC-parent-reopen-cascade holds for imported state, which
        arrives with no cascade to repair it). Imported
        events are preserved verbatim, and import appends exactly one
        `project.imported` event, stamped with `actor`, that never re-stamps
        per-record creation events. `actor` is the identity performing the
        import and must reference an identity contained in the export's
        `identities` — an actor existing only on the server would break the
        round-trip guarantee, since re-exporting would then contain an
        identity absent from the original. A round-trip comparison
        (AC-import-round-trip) excludes this appended import-audit event;
        re-exporting an imported project matches the original once that one
        event is discounted.
      tags: [projects]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
        - name: actor
          in: query
          required: true
          description: >-
            The identity performing the import, recorded as the actor on
            the appended `project.imported` event. Must resolve to an
            identity contained in the export's `identities` — the
            round-trip guarantee requires every identity in a re-export to
            trace back to the original, so an identity that exists only on
            this server is not a valid actor here. Such an actor is
            rejected with 400 (AC-import-actor).
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectExport'
      responses:
        '201':
          description: Project reconstructed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'

  # ------------------------------------------------------------------ issues
...
  /comments:
    post:
      operationId: createComment
      summary: Comment on an issue, doc version, or review
      description: >-
        Stores a comment with author and timestamp, anchored to exactly one of
        issue, doc_version, or review (AC-comment-create). A comment replying to
        another nests under its parent with unbounded thread depth
        (AC-comment-threading, AC-review-threads, AC-docweb-threads). Invoked by
        the comment action in the issue, document, and review views (ACT-comment).
      tags: [comments]
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewComment'
      responses:
        '201':
          description: Comment created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Comment'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
    get:
      operationId: listComments
      summary: List comments anchored to an issue, doc version, or review
      tags: [comments]
      parameters:
        - name: issue
          in: query
          schema: { type: string, format: uuid }
        - name: doc_version
          in: query
          schema: { type: string, format: uuid }
        - name: review
          in: query
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Comments, threaded via parent.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Comment' }

  # -------------------------------------------------------------- documents
...
components:
  parameters:
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: >-
        Client-supplied key that makes this mutating operation safe to retry
        (CON-idempotent-mutations): replaying the same key returns the
        original response verbatim and has no second effect — no duplicate
        resource, no second state transition, no second side effect.
      schema: { type: string }
...
  responses:
    BadRequest:
      description: >-
        The request is invalid — e.g. an actor or assignee identity id that
        names no known identity (AC-identity-referenced), or a status outside
        the five defined values (AC-status-invalid). The API deals only in
        identity ids; resolving a human-typed handle to an id is the CLI's job.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: The referenced resource does not exist.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: >-
        The request conflicts with current state — e.g. closing without an
        approved review, a blocking or ancestry cycle, a colliding import, a
        write against an archived project, or a review-anchored comment or
        verdict naming a stale revision (superseded by a resubmission since
        the reviewer last saw it) — rejected atomically, never applied to
        content the reviewer did not see.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
...
    Error:
      type: object
      required: [message, code]
      allOf:
        - anyOf:
            - not:
                type: object
                properties:
                  code: { type: string, enum: [unique-violation] }
                required: [code]
            - type: object
              required: [conflicts]
              properties:
                conflicts: { type: array, minItems: 1, items: {} }
          description: A unique-violation always names at least one colliding resource in conflicts — a schema-valid response can never omit what clients need to act on the collision.
      properties:
        message:
          type: string
          description: Human-readable explanation of the rejection.
        code:
          type: string
          enum: [relation-exists, ancestry-cycle, blocking-cycle,
                 project-archived, expected-status-mismatch,
                 expected-revision-mismatch, expected-verdict-event-mismatch,
                 expected-base-mismatch, expected-default-head-mismatch,
                 expected-subtree-revision-mismatch, review-close-used,
                 review-consumed, review-not-approved, open-children,
                 missing-approval, uuid-collision, malformed-import,
                 unique-violation, bad-request, not-found]
          description: >-
            Stable machine-readable cause, for deterministic client
            classification — a closed enum, so schema-valid responses cannot
            invent values. unique-violation covers every uniqueness
            constraint (duplicate identity handles, template names, label
            names), with the colliding resource named in conflicts. Clients distinguishing transient races from
            permanent conditions (e.g. kriya's parenting reconciliation) rely
            on this field, never on message text.
        conflicts:
          type: array
          description: Ids of the resources this request conflicts with, if applicable.
          items: { type: string }
    IssueStatus:
      type: string
      enum: [open, in-progress, blocked, deferred, complete]
    RelationKind:
      type: string
      enum: [parent_of, blocks]
    IdentityKind:
      type: string
      enum: [human, agent]
    ReviewState:
      type: string
      enum: [open, changes-requested, approved]
...
    Issue:
      type: object
      additionalProperties: false
      required: [id, number, title, status, project, created, updated, subtree_revision]
      properties:
        id: { type: string, format: uuid }
        number:
          type: integer
          description: Project-scoped display alias (e.g. SUT-42); never canonical identity.
        title: { type: string }
        body: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/IssueStatus' }
        project: { type: string, format: uuid }
        assignee: { type: string, format: uuid, nullable: true }
        labels:
          type: array
          items: { $ref: '#/components/schemas/Label' }
        created: { type: string, format: date-time }
        updated: { type: string, format: date-time }
        subtree_revision:
          type: integer
          description: Incremented atomically on the issue and every ancestor by each status transition, child attachment, or child removal in its subtree — removal increments the former parent's chain.
...
    Label:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        color: { type: string, nullable: true }
...
    IssueRelation:
      type: object
      required: [id, from, to, kind]
      properties:
        id: { type: string, format: uuid }
        from: { type: string, format: uuid }
        to: { type: string, format: uuid }
        kind: { $ref: '#/components/schemas/RelationKind' }
...
    Comment:
      type: object
      description: >-
        Polymorphic and threaded — anchors to exactly one of issue, doc_version,
        or review, with review_revision required exactly when review is set.
        The invariant is schema-encoded via oneOf, the same technique the
        Review response schema uses to encode its deliverable oneOf under
        the omit-when-absent convention — applied identically here, for
        responses and for comments carried in ProjectExport imports alike.
      required: [id, author, body, created]
      properties:
        id: { type: string, format: uuid }
        issue: { type: string, format: uuid }
        doc_version: { type: string, format: uuid }
        review: { type: string, format: uuid }
        parent: { type: string, format: uuid }
        anchor:
          type: string
          description: Optional span/block locator within the anchored artifact.
        review_revision:
          type: integer
          description: >-
            For review-anchored comments, the review revision the comment
            was made against; stamped from the submitted value; preserved
            verbatim on import. Required when review is set, absent otherwise.
        author: { type: string, format: uuid }
        body: { type: string }
        created: { type: string, format: date-time }
      oneOf:
        - required: [issue]
          not:
            anyOf:
              - required: [doc_version]
              - required: [review]
              - required: [review_revision]
        - required: [doc_version]
          not:
            anyOf:
              - required: [issue]
              - required: [review]
              - required: [review_revision]
        - required: [review, review_revision]
          not:
            anyOf:
              - required: [issue]
              - required: [doc_version]
    NewComment:
      type: object
      description: >-
        Exactly one of issue, doc_version, or review must be set. When
        anchored to a review, review_revision is required — the revision
        the reviewer saw; a value naming a stale revision (superseded by a
        resubmission) is rejected (AC-review-stale-guard).
      required: [author, body]
      properties:
        issue: { type: string, format: uuid }
        doc_version: { type: string, format: uuid }
        review: { type: string, format: uuid }
        review_revision:
          type: integer
          description: >-
            Required when review is set: the review revision the reviewer
            saw. Rejected if it names a revision superseded by a
            resubmission since (AC-review-stale-guard).
        parent: { type: string, format: uuid }
        anchor: { type: string }
        author: { type: string, format: uuid }
        body: { type: string }
      oneOf:
        - required: [issue]
          not:
            anyOf:
              - required: [doc_version]
              - required: [review]
              - required: [review_revision]
        - required: [doc_version]
          not:
            anyOf:
              - required: [issue]
              - required: [review]
              - required: [review_revision]
        - required: [review, review_revision]
          not:
            anyOf:
              - required: [issue]
              - required: [doc_version]
    Project:
      type: object
      required: [id, key, name]
      properties:
        id: { type: string, format: uuid }
        key: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        repo_path: { type: string, nullable: true }
        default_branch:
          type: string
          nullable: true
          description: >-
            Diff base for code deliverables (AC-review-web). Defaults to
            "main" when not set at creation.
        archived_at:
          type: string
          format: date-time
          nullable: true
          description: Set means archived (hidden, read-only).
...
    Document:
      type: object
      required: [id, project, title]
      properties:
        id: { type: string, format: uuid }
        project: { type: string, format: uuid }
        issue: { type: string, format: uuid, nullable: true }
        title: { type: string }
        current_version: { type: string, format: uuid, nullable: true }
...
    DocVersion:
      type: object
      required: [id, document, number, content, author, created]
      properties:
        id: { type: string, format: uuid }
        document: { type: string, format: uuid }
        number: { type: integer }
        content: { type: string }
        author: { type: string, format: uuid }
        created: { type: string, format: date-time }
...
    Thread:
      type: object
      description: >-
        An imported agent transcript, tied to a project or an issue. At
        least one of project or issue is always set — the invariant is
        schema-encoded via anyOf under the omit-when-absent convention (an
        absent side is omitted, never sent as null) — and applies equally
        to imported threads carried in ProjectExport/importProject payloads,
        not just to threads created through importThread/setThreadAnchor.
      required: [id, title, transcript, imported_at]
      properties:
        id: { type: string, format: uuid }
        title: { type: string }
        transcript:
          description: Arbitrary JSON transcript content, preserved verbatim.
        session:
          type: string
          nullable: true
          description: The instance this transcript came from; joins with Review.session.
        project: { type: string, format: uuid }
        issue: { type: string, format: uuid }
        imported_at: { type: string, format: date-time }
      anyOf:
        - required: [project]
        - required: [issue]
...
    Identity:
      type: object
      required: [id, handle, kind]
      properties:
        id: { type: string, format: uuid }
        handle: { type: string }
        kind: { $ref: '#/components/schemas/IdentityKind' }
        display_name: { type: string, nullable: true }
...
    Event:
      description: >-
        Discriminated by kind: `issue.relation-removed` events REQUIRE a
        RelationRemovedPayload; every other kind uses the general variant.
      oneOf:
        - $ref: '#/components/schemas/RelationRemovedEvent'
        - $ref: '#/components/schemas/GeneralEvent'
...
    Review:
      type: object
      description: >-
        A deliverable prepared for human approval — exactly one of a code
        deliverable (a branch pinned at an immutable commit; branch is
        display metadata, the commit is what approval covers) or doc_version
        (immutable, so approval pins a version). Absent fields are omitted,
        never sent as null — this applies to every optional field on this
        schema. `revision` starts at 1 and increments on each resubmission
        (AC-review-rework); a verdict and its approval apply to the review's
        current revision, and for a code deliverable that means the commit
        pinned as of that revision. These top-level fields mirror the latest
        submission; `submissions` holds the immutable history.
      required: [id, issue, author, state, revision, submissions, created]
      allOf:
        - anyOf:
            - required: [consumed, consumed_revision]
            - not:
                anyOf:
                  - required: [consumed]
                  - required: [consumed_revision]
          description: consumed and consumed_revision appear together or not at all — a half-consumed review is malformed, in responses and importProject payloads alike.
        - anyOf:
            - required: [close_used, consumed, consumed_revision]
            - not:
                required: [close_used]
          description: close_used implies both consumption fields. Beyond schema reach but enforced at import — EVERY consumed review, close-used or not, must be in state approved with consumed_revision equal to its current revision; importProject rejects violations as malformed, so a consumed review's verdict can never be rewritten through import and no state impossible through the API can be imported.
        - anyOf:
            - not:
                properties:
                  state: { enum: [approved, changes-requested] }
                required: [state]
            - required: [latest_verdict_event]
          description: Verdict-bearing states require latest_verdict_event — without it a review could never be closed, consumed, or resubmitted. Import additionally validates that it references the review's actual latest verdict event.
      properties:
        id: { type: string, format: uuid }
        issue: { type: string, format: uuid }
        branch: { type: string }
        commit:
          type: string
          pattern: "^[0-9a-f]{40}([0-9a-f]{24})?$"
          description: >-
            Canonical full object id — never an abbreviation. Immutable,
            pinned at creation/resubmission; branch is display metadata,
            this is what approval covers.
        doc_version: { type: string, format: uuid }
        session: { type: string }
        summary: { type: string }
        author: { type: string, format: uuid }
        state: { $ref: '#/components/schemas/ReviewState' }
        revision:
          type: integer
          description: >-
            1 on creation, +1 per resubmission; server-managed. Feedback
            left on an earlier revision stays associated with that
            revision — it does not move forward.
        latest_verdict_event:
          type: string
          format: uuid
          description: Id of the review's most recent verdict event; replaced by every verdict change. Delayed close, consume, and resubmit requests name it and are rejected when it moved — the verdict-ABA fence. Round-trips through export/import.
        consumed:
          type: string
          format: date-time
          description: >-
            Stamped by atomic approval consumption (AC-review-consume);
            present only once consumed. After this stamp the review's
            verdict can no longer change, and the stamp round-trips
            through export/import — where it is valid only on a review in
            state approved whose consumed_revision equals its current
            revision, close-used or not.
        consumed_revision:
          type: integer
          description: The revision whose approval was consumed; present exactly when `consumed` is.
        close_used:
          type: string
          format: date-time
          description: Stamped atomically by the completion this review authorized; a spent review can never authorize another close. Round-trips through export/import, where it is valid only on an approved review carrying both consumption fields with consumed_revision equal to the current revision.
        submissions:
          type: array
          description: >-
            One immutable ReviewSubmission per revision, oldest first —
            exactly what each revision submitted, preserved even after
            later resubmissions replace the top-level deliverable fields.
            Accepted verbatim in importProject payloads so revision history
            round-trips.
          items: { $ref: '#/components/schemas/ReviewSubmission' }
        created: { type: string, format: date-time }
      oneOf:
        - required: [branch, commit]
          not:
            required: [doc_version]
        - required: [doc_version]
          not:
            anyOf:
              - required: [branch]
              - required: [commit]
...
    ReviewExport:
      description: >-
        A Review as carried by ProjectExport: identical to Review with
        submissions strengthened to ReviewSubmissionExport and required
        to be NON-EMPTY, so exported review content always round-trips
        losslessly. importProject additionally validates — rejecting the
        payload as malformed on violation — that submission revisions
        are complete and contiguous from 1 through the review's current
        revision, and that the latest submission's deliverable matches
        the review's top-level deliverable fields, so no contract-valid
        export can smuggle a content-free or history-torn code review.
      allOf:
        - $ref: '#/components/schemas/Review'
        - type: object
          required: [submissions]
          properties:
            submissions:
              type: array
              minItems: 1
              items: { $ref: '#/components/schemas/ReviewSubmissionExport' }
...
    DocumentExport:
      type: object
      required: [document, versions]
      properties:
        document: { $ref: '#/components/schemas/Document' }
        versions:
          type: array
          items: { $ref: '#/components/schemas/DocVersion' }
    ProjectExport:
      type: object
      required:
        [project, identities, issues, comments, labels, issue_relations, documents, threads, reviews, events]
      properties:
        project: { $ref: '#/components/schemas/Project' }
        identities:
          type: array
          description: >-
            Every identity referenced by this export's records — authors,
            actors, assignees — so import recreates them instead of inventing
            or dropping references.
          items: { $ref: '#/components/schemas/Identity' }
        issues:
          type: array
          items: { $ref: '#/components/schemas/Issue' }
        comments:
          type: array
          items: { $ref: '#/components/schemas/Comment' }
        labels:
          type: array
          items: { $ref: '#/components/schemas/Label' }
        issue_relations:
          type: array
          items: { $ref: '#/components/schemas/IssueRelation' }
        documents:
          type: array
          items: { $ref: '#/components/schemas/DocumentExport' }
        threads:
          type: array
          items: { $ref: '#/components/schemas/Thread' }
        reviews:
          type: array
          items: { $ref: '#/components/schemas/ReviewExport' }
        events:
          type: array
          items: { $ref: '#/components/schemas/Event' }
...